+++ Src/Evora/Engine/floor.cpp
@@ -2,8 +2,8 @@
 
 namespace model
 {
-	const std::vector<int> floor::FLOOR_SCORE = { -1, -1, -2, -2, -2, -3, -3 };
-	
+	const std::vector<int> floor::FLOOR_SCORE = {-1, -1, -2, -2, -2, -3, -3};
+
 	std::vector<tile>::const_iterator floor::begin() const
 	{
 		return m_tiles.cbegin();
@@ -36,7 +36,7 @@ namespace model
 
 	int floor::move_to_lid(lid& lid)
 	{
-		int size = m_tiles.size();
+		int size = static_cast<int>(m_tiles.size());
 		for (auto tile : m_tiles)
 		{
 			lid.add_tile(tile);
@@ -47,9 +47,9 @@ namespace model
 
 	int floor::handle_starter_tile()
 	{
-		for(int i = 0; i < m_tiles.size(); ++i)
+		for (int i = 0; i < m_tiles.size(); ++i)
 		{
-			if(m_tiles[i] == tile::starter)
+			if (m_tiles[i] == tile::starter)
 			{
 				m_tiles.erase(m_tiles.begin() + i);
 				return i;
